Crate flate

Unstable (rustc_private)

Simple DEFLATE-based compression. This is a wrapper around the miniz library, which is a one-file pure-C implementation of zlib.

Structs

Bytes 	 [Unstable]

Error 	 [Unstable]

Functions

deflate_bytes 	 [Unstable]
Compress a buffer, without writing any sort of header on the output.

deflate_bytes_zlib 	 [Unstable]
Compress a buffer, using a header that zlib can understand.

inflate_bytes 	 [Unstable]
Decompress a buffer, without parsing any sort of header on the input.

inflate_bytes_zlib 	 [Unstable]
Decompress a buffer that starts with a zlib header.